requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
pydantic>=2.7.4
pydantic-settings>=2.3.4
authlib>=1.3.1
streamlit>=1.35.0
apscheduler>=3.10.4
openai>=1.35.3
faiss-cpu>=1.8.0
ddgs>=5.0.0
python-dotenv>=1.0.1
httpx>=0.27.0
email-validator>=2.1.1
itsdangerous>=2.2.0
